requirements.txt
asgiref==3.5.1
attrs==21.4.0
blessings==1.7
bpython==0.22.1
build==0.8.0
certifi==2022.6.15
charset-normalizer==2.0.12
cli-helpers==2.2.1
click==7.1.2
colored==1.4.3
configobj==5.0.6
curtsies==0.3.10
cwcwidth==0.1.6
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-filter==21.1
django-import-export==2.8.0
django-simple-history==3.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
greenlet==1.1.2
idna==3.3
importlib-resources==5.8.0
iniconfig==1.1.1
iredis==1.11.1
isort==5.10.1
lazy-object-proxy==1.7.1
MarkupPy==1.14
mccabe==0.7.0
mistune==2.0.2
numpy==1.22.4
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pendulum==2.1.2
pep517==0.12.0
pgcli==3.4.1
pgspecial==1.13.1
Pillow==9.1.1
pluggy==1.0.0
prompt-toolkit==3.0.29
psutil==5.9.1
psycopg2==2.9.3
py==1.11.0
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.1
pytzdata==2020.1
pyxdg==0.28
PyYAML==6.0
redis==3.5.3
requests==2.28.0
sentry-sdk==1.5.12
setproctitle==1.2.3
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tabulate==0.8.9
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
vizex==2.1.0
wcwidth==0.1.9
xlrd==2.0.1
xlwt==1.3.0
